Fix test breakage due to Caused by: java.lang.NoSuchMethodException: com.zutubi.pulse.acceptance.pages.browse.BuildSummaryPage.<init>(com.thoughtworks.selenium.Selenium, com.zutubi.pulse.master.webwork.Urls, java.lang.String, java.lang.Integer)

+++ com.zutubi.pulse.acceptance/src/test/com/zutubi/pulse/acceptance/MavenAcceptanceTest.java
@@ -70,7 +70,7 @@ public class MavenAcceptanceTest extends SeleniumTestBase
         loginAsAdmin();
         createMaven2Project();
 
-        int buildNumber = runBuild(random);
+        long buildNumber = runBuild(random);
 
         // We expect the summary page to report that 1 test passed.
         BuildSummaryPage summaryPage = browser.openAndWaitFor(BuildSummaryPage.class, random, buildNumber);
